ai_xau_bot/requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
ta==0.11.0
plotly==5.18.0
joblib==1.3.2
python-dotenv==1.0.0
